Prepare taxonomy block for react query v5. This also fixes a state error that occurred when switching versions
